add deployment create,get function in e2e fromework,delete createpvc.go, add lvmpvc.go
